Introduced a second phase of equipment modifications
The "We need the best quality for our men" choice provides
absolute benefits that diminish with time and relative penalties
that stay the same as the game progresses. Additional bonuses are
now provided in 1943 for armor, and once turbojet engine is researched
for aircraft. This makes the choice more acceptable, especially in light
of the heavy TC penalties.